test(ipc): co-locate diarizer tests and add permissions tests (#711)
Move the six diarizer-command tests from commands/mod.rs into a
#[cfg(test)] mod tests block at the bottom of commands/diarizer.rs,
where they belong alongside the code they exercise. Motivation:
commands/mod.rs is already sizeable; diarizer.rs has testable seams
(download_diarizer_model_inner, swap_diarizer_after_download,
DiarizerDownloadDeps) but no test module — a gap that made coverage
invisible in file-level review.

Tests moved (unchanged logic, updated super:: references):
- swap_diarizer_after_download_err_leaves_slot_intact
- remove_diarizer_model_is_idempotent_when_file_missing
- remove_diarizer_model_persists_toggle_off
- download_diarizer_model_rejects_duplicate_concurrent_clicks
- download_diarizer_model_clears_cancel_handle_on_failure
Plus helpers: SwapSentinelDiarizer, remove_diarizer_model_test,
make_test_diarizer_model, build_download_deps.

New tests added to commands/permissions.rs:
- open_macos_privacy_pane_rejects_unknown_target (macOS-only):
  confirms the whitelist rejects arbitrary pane names.
- open_macos_privacy_pane_accepts_whitelisted_targets (macOS-only):
  pins the complete whitelist so a removed/misspelled key is a test
  failure rather than a silent runtime regression.
- diagnose_macos_permissions_non_macos_returns_empty_bundle_id_and_no_reset
  (non-macOS / Linux CI): confirms the non-macOS fallthrough returns
  bundle_id="" and can_reset=false so the Settings panel suppresses
  the Reset button on unsupported platforms.

Also removes the now-unused AppState import from commands/mod.rs tests.
