+++ src-tauri/src/ipc/commands/mod.rs
@@ -308,8 +308,6 @@ pub const UPDATE_CHECK_TTL: std::time::Duration = std::time::Duration::from_secs
 
 #[cfg(test)]
 mod tests {
-    use crate::ipc::AppState;
-
     use super::*;
 
     #[test]
@@ -574,69 +572,6 @@ mod tests {
         );
     }
 
-    /// Sentinel diarizer used by the swap-failure test below.
-    /// Different type from the `RecordingDiarizer` in
-    /// `diarization::tests` so we can use `Arc::ptr_eq` reliably
-    /// to confirm the *exact same* `Arc` survived the failed swap.
-    /// Gated alongside the test that uses it so `--no-default-features`
-    /// builds don't trip the dead-code lint.
-    #[cfg(feature = "diarization-onnx")]
-    struct SwapSentinelDiarizer;
-
-    #[cfg(feature = "diarization-onnx")]
-    impl crate::diarization::Diarize for SwapSentinelDiarizer {
-        fn label_utterances(
-            &self,
-            _utterances: &mut [crate::transcription::Utterance],
-            _audio_chunks: &[Vec<f32>],
-            _format: crate::audio::CaptureFormat,
-        ) {
-            // No-op; presence in the slot is the assertion.
-        }
-    }
-
-    #[cfg(feature = "diarization-onnx")]
-    #[test]
-    fn swap_diarizer_after_download_err_leaves_slot_intact() {
-        // Audit-2 gap: when the post-download model load fails
-        // (corrupt ONNX, SHA mismatch from `OnnxDiarizer::new`'s
-        // load-time verify, or feature compiled out), the slot
-        // must not be poisoned or replaced with a half-built
-        // diarizer. The catch path in `download_diarizer_model`
-        // also relies on this — if the slot got partially written
-        // on Err, a subsequent successful swap could pile on top
-        // of an indeterminate state.
-        //
-        // Test: build a slot with a sentinel diarizer; call swap
-        // with a tempfile whose contents won't match the
-        // wespeaker SHA (so `OnnxDiarizer::new` fails *before*
-        // any `slot.write()` happens); assert the slot still
-        // points at the exact same Arc via `Arc::ptr_eq`.
-        use std::io::Write;
-        let dir = tempfile::tempdir().expect("tempdir");
-        let path = dir.path().join("not-wespeaker.onnx");
-        let mut f = std::fs::File::create(&path).expect("create");
-        f.write_all(b"definitely not a wespeaker model")
-            .expect("write");
-        drop(f);
-
-        let sentinel: std::sync::Arc<dyn crate::diarization::Diarize> =
-            std::sync::Arc::new(SwapSentinelDiarizer);
-        let slot: crate::diarization::DiarizeSlot =
-            std::sync::Arc::new(std::sync::RwLock::new(std::sync::Arc::clone(&sentinel)));
-
-        let res = super::diarizer::swap_diarizer_after_download(&slot, &path);
-        assert!(res.is_err(), "swap should reject a non-wespeaker file");
-
-        // The slot still holds the sentinel — same Arc identity,
-        // not a clone or replacement.
-        let guard = slot.read().expect("slot read");
-        assert!(
-            std::sync::Arc::ptr_eq(&*guard, &sentinel),
-            "swap failure must not replace the slot's Arc"
-        );
-    }
-
     // ---- check_for_updates cache (#333) --------------------------------
 
     #[tokio::test]
@@ -714,260 +649,5 @@ mod tests {
     // History CSV export (#357 phase 3a) tests live in
     // `crate::ipc::commands::history` — moved alongside the helper
     // under #431.
-
-    // -- remove_diarizer_model (#351) ----------------------------------
-
-    #[tokio::test]
-    async fn remove_diarizer_model_is_idempotent_when_file_missing() {
-        // Removing when the file isn't present must succeed cleanly
-        // — covers the race where two `remove` calls fire (or the
-        // user deleted the file out of band before clicking
-        // Remove). Slot still gets reverted to a Noop-shaped
-        // diarizer either way so the in-memory state stays
-        // consistent. Mock state's models_dir is a fresh tempdir;
-        // the wespeaker file is not present.
-        let state = crate::ipc::tests::mock_state();
-        remove_diarizer_model_inner(&state)
-            .await
-            .expect("idempotent on missing file");
-        // The slot swap is exercised separately by
-        // `swap_diarizer_after_download_err_leaves_slot_intact` and
-        // friends; here we just pin that the call succeeded
-        // without panicking and the toggle persistence below
-        // landed.
-    }
-
-    #[tokio::test]
-    async fn remove_diarizer_model_persists_toggle_off() {
-        // The Speakers panel reads `diarization_enabled` to drive
-        // the toggle UI. Remove must clear the flag (in-memory
-        // atomic + persisted setting row) so a re-install lands
-        // in a consistent off-by-default state.
-        let state = crate::ipc::tests::mock_state();
-        // Set the toggle on first so the `remove` flip is observable.
-        state
-            .runtime_flags
-            .diarization_enabled
-            .store(true, std::sync::atomic::Ordering::Relaxed);
-        state
-            .settings
-            .set(crate::settings::keys::DIARIZATION_ENABLED, "true")
-            .await
-            .expect("seed settings");
-
-        remove_diarizer_model_inner(&state)
-            .await
-            .expect("remove ok");
-
-        assert!(
-            !state
-                .runtime_flags
-                .diarization_enabled
-                .load(std::sync::atomic::Ordering::Relaxed),
-            "atomic should flip to false"
-        );
-        let persisted = state
-            .settings
-            .get(crate::settings::keys::DIARIZATION_ENABLED)
-            .await
-            .expect("settings get");
-        assert_eq!(persisted.as_deref(), Some("false"));
-    }
-
-    /// Test-side wrapper that mirrors the IPC body — keeps the
-    /// `#[tauri::command]` shell out of the test path so we don't
-    /// need a `tauri::State<'_, AppState>` constructor.
-    async fn remove_diarizer_model_inner(state: &AppState) -> IpcResult<()> {
-        let model = crate::diarization::catalog::default_diarizer_model();
-        let path = state.models_dir.join(&model.filename);
-        match tokio::fs::remove_file(&path).await {
-            Ok(()) => {}
-            Err(e) if e.kind() == std::io::ErrorKind::NotFound => {}
-            Err(e) => {
-                return Err(IpcError::Internal(format!(
-                    "remove diarizer model {}: {e}",
-                    path.display()
-                )));
-            }
-        }
-        {
-            let mut slot = state
-                .diarize_slot
-                .write()
-                .unwrap_or_else(|e| e.into_inner());
-            *slot = std::sync::Arc::new(crate::diarization::NoopDiarizer);
-        }
-        state
-            .runtime_flags
-            .diarization_enabled
-            .store(false, std::sync::atomic::Ordering::Relaxed);
-        state
-            .settings
-            .set(crate::settings::keys::DIARIZATION_ENABLED, "false")
-            .await
-            .map_err(|e| IpcError::Settings(e.to_string()))?;
-        Ok(())
-    }
-
-    // ---- #315: download_diarizer_model_inner via EventEmitter ----
-
-    /// Build a synthetic diarizer-model entry pointing at a URL the
-    /// test wants the http path to hit. Used by the failure-cleanup
-    /// test to drive the download into the failure branch via an
-    /// unbindable port; SHA + filename are arbitrary because the
-    /// test asserts on cancel-handle cleanup, not on payload
-    /// content.
-    fn make_test_diarizer_model(url: &str) -> crate::diarization::catalog::DiarizerModelMetadata {
-        crate::diarization::catalog::DiarizerModelMetadata {
-            id: "wespeaker-test".into(),
-            display_name: "Wespeaker (test)".into(),
-            filename: "test_diarizer.onnx".into(),
-            size_mb: 1,
-            description: "test entry".into(),
-            download_url: url.into(),
-            sha256: "0".repeat(64),
-        }
-    }
-
-    fn build_download_deps(
-        emitter: std::sync::Arc<dyn crate::events::EventEmitter>,
-        downloads: std::sync::Arc<
-            std::sync::Mutex<
-                std::collections::HashMap<String, crate::transcription::download::CancelHandle>,
-            >,
-        >,
-        models_dir: std::path::PathBuf,
-    ) -> super::diarizer::DiarizerDownloadDeps {
-        super::diarizer::DiarizerDownloadDeps {
-            emitter,
-            downloads,
-            http: reqwest::Client::new(),
-            // Tests don't exercise the swap path; a NoopDiarizer
-            // slot is enough to satisfy the type. Even the
-            // failure-cleanup test bails before the
-            // swap_diarizer_after_download call.
-            diarize_slot: std::sync::Arc::new(std::sync::RwLock::new(std::sync::Arc::new(
-                crate::diarization::NoopDiarizer,
-            ))),
-            models_dir,
-        }
-    }
-
-    #[tokio::test]
-    async fn download_diarizer_model_rejects_duplicate_concurrent_clicks() {
-        // Pre-seed the downloads map with the diarizer id (as if a
-        // prior click had spawned a task). The second call must
-        // bail with `IpcError::Settings` and emit no events.
-        let downloads = std::sync::Arc::new(std::sync::Mutex::new(std::collections::HashMap::<
-            String,
-            crate::transcription::download::CancelHandle,
-        >::new()));
-        let model = make_test_diarizer_model("http://127.0.0.1:1/never-fetched");
-        downloads.lock().unwrap().insert(
-            model.id.clone(),
-            crate::transcription::download::CancelHandle::new(),
-        );
-
-        let recorder = crate::ipc::events::RecordingEventEmitter::new();
-        let emitter: std::sync::Arc<dyn crate::events::EventEmitter> =
-            std::sync::Arc::new(recorder.clone());
-
-        let tmp = tempfile::tempdir().unwrap();
-        let deps = build_download_deps(
-            emitter,
-            std::sync::Arc::clone(&downloads),
-            tmp.path().to_path_buf(),
-        );
-
-        let result = super::diarizer::download_diarizer_model_inner(deps, model.clone()).await;
-        match result {
-            Err(IpcError::Settings(msg)) => {
-                assert!(
-                    msg.contains("already downloading"),
-                    "expected duplicate-rejection message, got: {msg}"
-                );
-            }
-            other => panic!("expected IpcError::Settings, got: {other:?}"),
-        }
-
-        assert!(
-            recorder.events().is_empty(),
-            "duplicate rejection should not emit any events; got {:?}",
-            recorder.events()
-        );
-
-        // The pre-existing handle must still be in place; the
-        // rejection path should not have touched it (regression
-        // guard for "rejection accidentally clears the slot").
-        let still_present = downloads.lock().unwrap().contains_key(&model.id);
-        assert!(still_present, "pre-existing cancel handle was clobbered");
-    }
-
-    #[tokio::test]
-    async fn download_diarizer_model_clears_cancel_handle_on_failure() {
-        // Drive the download into the failure branch by pointing
-        // it at an unbindable port (127.0.0.1:1). reqwest will
-        // surface a connect error and the spawned task takes the
-        // `Err(e)` arm of the match, which must:
-        //   - remove its cancel-handle entry from `downloads`, AND
-        //   - emit `model:download-failed` with the chained error.
-        // Pre-#315 there was no test for this; the `try_state`
-        // hop in the cleanup made the path reachable only from a
-        // live Tauri runtime.
-        let downloads = std::sync::Arc::new(std::sync::Mutex::new(std::collections::HashMap::<
-            String,
-            crate::transcription::download::CancelHandle,
-        >::new()));
-        let recorder = crate::ipc::events::RecordingEventEmitter::new();
-        let emitter: std::sync::Arc<dyn crate::events::EventEmitter> =
-            std::sync::Arc::new(recorder.clone());
-
-        let tmp = tempfile::tempdir().unwrap();
-        let model = make_test_diarizer_model("http://127.0.0.1:1/will-fail");
-        let deps = build_download_deps(
-            emitter,
-            std::sync::Arc::clone(&downloads),
-            tmp.path().to_path_buf(),
-        );
-
-        super::diarizer::download_diarizer_model_inner(deps, model.clone())
-            .await
-            .expect("inner returns Ok before the spawn — the failure happens inside the task");
-
-        // Wait for the spawned task to finish. The connect error
-        // surfaces in single-digit ms locally; bound the wait at
-        // 5s with a polling loop so a CI hiccup doesn't hang.
-        let cleared = tokio::time::timeout(std::time::Duration::from_secs(5), async {
-            loop {
-                let still_in_flight = downloads.lock().unwrap().contains_key(&model.id);
-                if !still_in_flight {
-                    return true;
-                }
-                tokio::time::sleep(std::time::Duration::from_millis(20)).await;
-            }
-        })
-        .await
-        .unwrap_or(false);
-
-        assert!(
-            cleared,
-            "cancel handle should have been removed by the failure branch"
-        );
-
-        // Failure event must have fired with a non-empty message
-        // — the actual reqwest text varies by platform so we
-        // assert on the shape rather than the exact wording.
-        let failures = recorder.payloads_for("model:download-failed");
-        assert_eq!(
-            failures.len(),
-            1,
-            "exactly one failure event expected; got {failures:?}"
-        );
-        let payload = &failures[0];
-        assert_eq!(payload["id"], serde_json::Value::String(model.id.clone()));
-        let msg = payload["message"]
-            .as_str()
-            .expect("failure event should carry a message string");
-        assert!(!msg.is_empty(), "failure event message should be populated");
-    }
+    // Diarizer tests live in `crate::ipc::commands::diarizer` (#711).
 }
